refactor, add search view component
